+++ libs/designer-ui/src/lib/tokenpicker/plugins/InitializeTokenPickerExpressionHandler.tsx
@@ -1,28 +1,25 @@
-import { findChildNode } from '../../editor/base/utils/helper';
-import { TokenType } from '../../editor/models/parameter';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
 import type { LexicalCommand, NodeKey } from 'lexical';
-import { $getRoot, COMMAND_PRIORITY_EDITOR, createCommand } from 'lexical';
+import { COMMAND_PRIORITY_EDITOR, createCommand } from 'lexical';
+import type { OPEN_TOKEN_PICKER_PAYLOAD } from '../../editor/base/plugins/OpenTokenPicker';
 import { useEffect } from 'react';
 
-export const INITIALIZE_TOKENPICKER_EXPRESSION: LexicalCommand<string> = createCommand();
+export const INITIALIZE_TOKENPICKER_EXPRESSION: LexicalCommand<OPEN_TOKEN_PICKER_PAYLOAD> = createCommand();
 
 interface TokenPickerHandlerProps {
   handleInitializeExpression?: (s: string, n: NodeKey) => void;
 }
 
-export default function TokenPickerHandler({ handleInitializeExpression }: TokenPickerHandlerProps): null {
+export default function TokenPickerExpressionHandler({ handleInitializeExpression }: TokenPickerHandlerProps): null {
   const [editor] = useLexicalComposerContext();
 
   useEffect(() => {
-    return editor.registerCommand<string>(
+    return editor.registerCommand<OPEN_TOKEN_PICKER_PAYLOAD>(
       INITIALIZE_TOKENPICKER_EXPRESSION,
-      (payload: string) => {
-        const node = findChildNode($getRoot(), payload, TokenType.FX);
-        if (node?.token?.tokenType === TokenType.FX) {
-          handleInitializeExpression?.(node?.value ?? '', payload);
-        } else {
-          editor.focus();
+      (payload: OPEN_TOKEN_PICKER_PAYLOAD) => {
+        const { token, nodeKey } = payload;
+        if (token?.value) {
+          handleInitializeExpression?.(token.value, nodeKey);
         }
         return true;
       },