+++ libs/designer/src/lib/core/state/__test__/operation/operationMetadataSlice.spec.ts
@@ -22,6 +22,7 @@ describe('operationMetadataSlice', () => {
       actionMetadata: {},
       staticResults: {},
       repetitionInfos: {},
+      supportedChannels: {},
       errors: {},
       loadStatus: {
         nodesInitialized: false,