+++ libs/designer/src/lib/core/state/workflow/workflowSlice.ts
@@ -2,7 +2,7 @@ import constants from '../../../common/constants';
 import { updateNodeConnection } from '../../actions/bjsworkflow/connections';
 import { initializeGraphState } from '../../parsers/ParseReduxAction';
 import type { AddNodePayload } from '../../parsers/addNodeToWorkflow';
-import { addSwitchCaseToWorkflow, addNodeToWorkflow, addAgentConditionToWorkflow } from '../../parsers/addNodeToWorkflow';
+import { addSwitchCaseToWorkflow, addNodeToWorkflow, addAgentToolToWorkflow } from '../../parsers/addNodeToWorkflow';
 import type { DeleteNodePayload } from '../../parsers/deleteNodeFromWorkflow';
 import { deleteWorkflowNode, deleteNodeFromWorkflow } from '../../parsers/deleteNodeFromWorkflow';
 import type { WorkflowNode } from '../../parsers/models/workflowNode';
@@ -33,6 +33,8 @@ import {
   RUN_AFTER_STATUS,
   WORKFLOW_EDGE_TYPES,
   WORKFLOW_NODE_TYPES,
+  containsIdTag,
+  containsCaseTag,
 } from '@microsoft/logic-apps-shared';
 import type { MessageLevel } from '@microsoft/designer-ui';
 import { getDurationStringPanelMode } from '@microsoft/designer-ui';
@@ -255,8 +257,7 @@ export const workflowSlice = createSlice({
         return; // log exception
       }
       const { nodeId } = action.payload;
-      const graphId = getRecordEntry(state.nodesMetadata, nodeId)?.graphId ?? '';
-      const graph = getWorkflowNodeFromGraphState(state, graphId);
+      const graph = getWorkflowNodeFromGraphState(state, nodeId);
       if (!graph) {
         throw new Error('graph not set');
       }
@@ -276,6 +277,7 @@ export const workflowSlice = createSlice({
       }
       const { scopeRepetitionRunData, nodeId } = action.payload;
       const { tools = {} } = scopeRepetitionRunData ?? {};
+      const agentGraph = getWorkflowNodeFromGraphState(state, nodeId);
 
       Object.keys(tools).forEach((toolId: any) => {
         const nodeMetadata = getRecordEntry(state.nodesMetadata, toolId);
@@ -286,8 +288,9 @@ export const workflowSlice = createSlice({
           ...nodeMetadata,
           runData: {
             status: tools[toolId].status,
-            repetitionCount: tools[toolId].repetitions,
+            repetitionCount: tools[toolId].iterations,
           },
+          runIndex: 0,
         };
         state.nodesMetadata[toolId] = nodeData as NodeMetadata;
       });
@@ -307,11 +310,13 @@ export const workflowSlice = createSlice({
         ),
       };
       nodeMetadata.runData = nodeRunData as LogicAppsV2.WorkflowRunAction;
+      nodeMetadata.actionCount =
+        (agentGraph?.children ?? []).filter((node) => !containsIdTag(node.id) && !containsCaseTag(node.id))?.length ?? -1;
 
       LoggerService().log({
         level: LogEntryLevel.Verbose,
         area: 'Designer:Workflow Slice',
-        message: 'Update tools metadata',
+        message: 'Update agentic metadata',
         args: [action.payload],
       });
     },
@@ -456,6 +461,19 @@ export const workflowSlice = createSlice({
       };
       nodeMetadata.runData = nodeRunData as LogicAppsV2.WorkflowRunAction;
     },
+    setSubgraphRunData: (state: WorkflowState, action: PayloadAction<{ nodeId: string; runData: LogicAppsV2.RunRepetition[] }>) => {
+      const { nodeId, runData } = action.payload;
+      const nodeMetadata = getRecordEntry(state.nodesMetadata, nodeId);
+      if (!nodeMetadata) {
+        return;
+      }
+      const subgraph = runData.reduce((acc, run) => {
+        const nodeId = run.name;
+        acc[nodeId] = run.properties;
+        return acc;
+      }, {} as any);
+      nodeMetadata.subgraphRunData = subgraph;
+    },
     setRunDataInputOutputs: (
       state: WorkflowState,
       action: PayloadAction<{ nodeId: string; inputs: BoundParameters; outputs: BoundParameters }>
@@ -491,17 +509,17 @@ export const workflowSlice = createSlice({
         args: [action.payload],
       });
     },
-    addAgentCondition: (state: WorkflowState, action: PayloadAction<{ conditionId: string; nodeId: string }>) => {
+    addAgentCondition: (state: WorkflowState, action: PayloadAction<{ toolId: string; nodeId: string }>) => {
       if (!state.graph) {
         return; // log exception
       }
-      const { conditionId, nodeId } = action.payload;
+      const { toolId, nodeId } = action.payload;
       const graphId = getRecordEntry(state.nodesMetadata, nodeId)?.graphId ?? '';
       const node = getWorkflowNodeFromGraphState(state, graphId);
       if (!node) {
         throw new Error('node not set');
       }
-      addAgentConditionToWorkflow(conditionId, node, state.nodesMetadata, state);
+      addAgentToolToWorkflow(toolId, node, state.nodesMetadata, state);
 
       LoggerService().log({
         level: LogEntryLevel.Verbose,
@@ -739,6 +757,7 @@ export const {
   replaceId,
   setRunIndex,
   setRepetitionRunData,
+  setSubgraphRunData,
   setIsWorkflowDirty,
   setHostErrorMessages,
   setRunDataInputOutputs,