+++ libs/logic-apps-shared/src/designer-client-services/lib/run.ts
@@ -14,7 +14,10 @@ export interface IRunService {
     status?: string
   ): Promise<{ value: LogicAppsV2.RunRepetition[] }>;
   getAgentRepetition(action: { nodeId: string; runId: string | undefined }, repetitionId: string): Promise<LogicAppsV2.RunRepetition>;
+  getAgentActionsRepetition(action: { nodeId: string; runId: string | undefined }, repetitionId: string): Promise<any>;
+  getMoreAgentActionsRepetition(continuationToken: string): Promise<any>;
   getRepetition(action: { nodeId: string; runId: string | undefined }, repetitionId: string): Promise<LogicAppsV2.RunRepetition>;
+  getChatHistory(action: { nodeId: string; runId: string | undefined }): Promise<any>;
 }
 
 let service: IRunService;