+++ libs/logic-apps-shared/src/utils/src/lib/models/index.ts
@@ -19,6 +19,7 @@ export * from './uiInteractionData';
 export * from './topLevelDropdownMenuItem';
 export * from './token';
 export * from './logicAppsV2Integration';
+export * from './workflow';
 
 import * as LogicApps from './logicApps';
 import * as LogicAppsV2 from './logicAppsV2';